+++ composeApp/src/commonMain/kotlin/ai/mcpdirect/studio/app/model/OpenAPIServer.kt
@@ -1,0 +1,11 @@
+package ai.mcpdirect.studio.app.model
+
+import ai.mcpdirect.studio.app.model.aitool.AIPortToolMaker
+import kotlinx.serialization.Serializable
+
+@Serializable
+class OpenAPIServer: AIPortToolMaker() {
+    var url: String? = null
+
+    var securities: Map<String, OpenAPIServerConfig.Security>? = null
+}