+++ composeApp/src/commonMain/kotlin/ai/mcpdirect/studio/app/tool/MCPToolsScreen.kt
@@ -1,29 +1,21 @@
 package ai.mcpdirect.studio.app.tool
 
-import ai.mcpdirect.studio.app.auth.authViewModel
 import ai.mcpdirect.studio.app.compose.StudioActionBar
 import ai.mcpdirect.studio.app.compose.StudioCard
-import ai.mcpdirect.studio.app.compose.StudioIcon
-import ai.mcpdirect.studio.app.compose.StudioListItem
 import ai.mcpdirect.studio.app.compose.TooltipIconButton
-import ai.mcpdirect.studio.app.compose.TooltipText
 import ai.mcpdirect.studio.app.generalViewModel
-import ai.mcpdirect.studio.app.model.account.AIPortTeam
-import ai.mcpdirect.studio.app.model.aitool.AIPortToolMaker
 import ai.mcpdirect.studio.app.virtualmcp.VirtualToolMakerDetailView
 import ai.mcpdirect.studio.app.virtualmcp.VirtualToolMakerListView
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
-import androidx.compose.material3.*
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.SecondaryTabRow
+import androidx.compose.material3.Tab
+import androidx.compose.material3.Text
 import androidx.compose.runtime.*
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
-import mcpdirectstudioapp.composeapp.generated.resources.*
-import org.jetbrains.compose.resources.painterResource
+import mcpdirectstudioapp.composeapp.generated.resources.Res
+import mcpdirectstudioapp.composeapp.generated.resources.refresh
 
 private enum class MCPToolsScreenDialog {
     NONE,
@@ -96,90 +88,90 @@ fun MCPToolsScreen() {
         }
     }
 }
-
-@Composable
-fun MCPServerList(
-    toolListViewModel: ToolListViewModel,
-    modifier: Modifier = Modifier
-){
-    LaunchedEffect(null){
-        generalViewModel.refreshToolMakers()
-    }
-    val viewModel = mcpToolsViewModel
-    LazyColumn(modifier=modifier) {
-        items(generalViewModel.toolMakers){
-            val me = it.userId== authViewModel.user.id
-            val team: AIPortTeam? = generalViewModel.team(it.teamId)
-            StudioListItem(
-                selected = viewModel.toolMaker?.id == it.id,
-                modifier = Modifier.clickable{
-                    viewModel.toolMaker(it)
-                    toolListViewModel.toolMaker(it)
-                },
-                overlineContent = {
-                    Row {
-                        StudioIcon(
-                            Res.drawable.person,
-                            "Owner",
-                            modifier = Modifier.size(18.dp).padding(end = 4.dp)
-                        )
-                        if(me) Text("Me")
-                        else if(team!=null) TooltipText(
-                            team.ownerName,
-                            contentDescription = team.ownerAccount,
-                            overflow = TextOverflow.Ellipsis
-                        )
-                    }
-                },
-                headlineContent = { Text(
-                    it.name,
-                    maxLines = 1,
-                    overflow = TextOverflow.Ellipsis) },
-                supportingContent = {
-                    if(me) Row{
-                        StudioIcon(
-                            Res.drawable.graph_5,
-                            "From My Studio",
-                            MaterialTheme.colorScheme.primary,
-                            modifier = Modifier.size(20.dp).padding(end = 4.dp)
-                        )
-                        Text(
-                            if(it.type== AIPortToolMaker.TYPE_VIRTUAL)
-                                "Virtual MCP"
-                            else it.agentName,
-                            color = MaterialTheme.colorScheme.primary,
-                            style=MaterialTheme.typography.bodySmall
-                        )
-
-                    } else if(team!=null) Row{
-                        StudioIcon(
-                            Res.drawable.groups,
-                            "From Team",
-                            MaterialTheme.colorScheme.secondary,
-                            modifier = Modifier.size(20.dp).padding(end = 4.dp)
-                        )
-                        Text(
-                            team.name,
-                            color = MaterialTheme.colorScheme.secondary,
-                            style=MaterialTheme.typography.bodySmall)
-                    }
-
-                },
-                trailingContent = {
-                    if(viewModel.toolMaker!=null&&viewModel.toolMaker!!.id==it.id)
-                        Icon(painterResource(Res.drawable.keyboard_arrow_right),
-                            contentDescription = "Current Tool Maker")
-                },
-                colors = ListItemDefaults.colors(
-                    containerColor = if(viewModel.toolMaker==it)
-                        MaterialTheme.colorScheme.surfaceContainer
-                    else Color.Transparent
-                )
-            )
-            HorizontalDivider()
-        }
-    }
-}
+//
+//@Composable
+//fun MCPServerList(
+//    toolListViewModel: ToolListViewModel,
+//    modifier: Modifier = Modifier
+//){
+//    LaunchedEffect(null){
+//        generalViewModel.refreshToolMakers()
+//    }
+//    val viewModel = mcpToolsViewModel
+//    LazyColumn(modifier=modifier) {
+//        items(generalViewModel.toolMakers){
+//            val me = it.userId== authViewModel.user.id
+//            val team: AIPortTeam? = generalViewModel.team(it.teamId)
+//            StudioListItem(
+//                selected = viewModel.toolMaker?.id == it.id,
+//                modifier = Modifier.clickable{
+//                    viewModel.toolMaker(it)
+//                    toolListViewModel.toolMaker(it)
+//                },
+//                overlineContent = {
+//                    Row {
+//                        StudioIcon(
+//                            Res.drawable.person,
+//                            "Owner",
+//                            modifier = Modifier.size(18.dp).padding(end = 4.dp)
+//                        )
+//                        if(me) Text("Me")
+//                        else if(team!=null) TooltipText(
+//                            team.ownerName,
+//                            contentDescription = team.ownerAccount,
+//                            overflow = TextOverflow.Ellipsis
+//                        )
+//                    }
+//                },
+//                headlineContent = { Text(
+//                    it.name,
+//                    maxLines = 1,
+//                    overflow = TextOverflow.Ellipsis) },
+//                supportingContent = {
+//                    if(me) Row{
+//                        StudioIcon(
+//                            Res.drawable.graph_5,
+//                            "From My Studio",
+//                            MaterialTheme.colorScheme.primary,
+//                            modifier = Modifier.size(20.dp).padding(end = 4.dp)
+//                        )
+//                        Text(
+//                            if(it.type== AIPortToolMaker.TYPE_VIRTUAL)
+//                                "Virtual MCP"
+//                            else it.agentName,
+//                            color = MaterialTheme.colorScheme.primary,
+//                            style=MaterialTheme.typography.bodySmall
+//                        )
+//
+//                    } else if(team!=null) Row{
+//                        StudioIcon(
+//                            Res.drawable.groups,
+//                            "From Team",
+//                            MaterialTheme.colorScheme.secondary,
+//                            modifier = Modifier.size(20.dp).padding(end = 4.dp)
+//                        )
+//                        Text(
+//                            team.name,
+//                            color = MaterialTheme.colorScheme.secondary,
+//                            style=MaterialTheme.typography.bodySmall)
+//                    }
+//
+//                },
+//                trailingContent = {
+//                    if(viewModel.toolMaker!=null&&viewModel.toolMaker!!.id==it.id)
+//                        Icon(painterResource(Res.drawable.keyboard_arrow_right),
+//                            contentDescription = "Current Tool Maker")
+//                },
+//                colors = ListItemDefaults.colors(
+//                    containerColor = if(viewModel.toolMaker==it)
+//                        MaterialTheme.colorScheme.surfaceContainer
+//                    else Color.Transparent
+//                )
+//            )
+//            HorizontalDivider()
+//        }
+//    }
+//}
 
 //@Composable
 //fun MCPServerItem(